refactor: moved the network status logics in a hook
